test(highlight): add failing test for offset calc
When the Jest rootDir is not the same as the package's root directory
(i.e., a monorepo scenario with multiple local Jest configs) the
highlight function incorrectly calculates the offset for highlighting file
name matches, resulting in a highlight in the wrong place.
This only occurs when filePath is not truncated.